Removed getTemplate function
This probably will break some code around :)

The interface of ITemplating has changed a little.  Its method .template
returns always the relation object, which can be None (as before, so continue
testing its value :), but now a .template_object method is provided, and this
method returns the real template, even if there's no relation object set (in
which case a NullTemplate is returned).  This simplifies the logic of client
code, which now can rely on ITemplating and be sure to get a valid template
(although it can be a null template).

This has two implications:
- No need for getTemplate: code relying on it will break, because it was
  removed.  Please, rely on ITemplating and .template_object.
- No need for catching MissingTemplateError, since there's always a meaningful
  template (and also code relying on this exception will break because it was
  removed too).