Fix links when building a page
All links of a page built with templates must have their URLs changed to
absolute ones.

The template content itself is generally authored using the current context
URL as a base for all of its links.  The same generally happens with the
content which will receive the template.  So, all links of the template and
the content must be independently absolutized before they mix in together.
This ensures that all links are valid in the final page built.