Revert "Possible bug with grok..."
This reverts commit 73b07fb98940d9f9a61df4ae6163e50e490119aa.

Although there *is* a bug in grok's scanning code, the solution of the
previous commit does not solve the problem.

The bug is in martian.scan (martian version 0.11.2).  Not sure if this bug
happens in newer martian versions.

martian.scan doesn't treat modules whose path ends with __init__.pyo as
packages, at leat in the way grok uses it.  It works, however, if the path ends
with __init__pyc.

It turns out that PYTHONOPTIMIZE must not be set with that specific version of
martian.

For now, no code in this package will fix that, so we're reverting that
commit, since it worked only by coincidence.  The coincidence was that a newer
source version as used when we checked out this package in production, and a
new module was compiled by Python, and martian correctly did its magic and
picked up our components.  A subsequent restart in the Zope instance, where a
.pyo package is used due to previous compilation (and optimization) make it
fail and our components don't show up, causing ComponentLookupErrors.

Disabling PYTHONOPTIMIZE works just fine.